requirements.txt
appdirs==1.4.2
beautifulsoup4==4.5.3
Django==1.10.5
django-cors-headers==2.0.2
django-filter==1.0.1
django-modelcluster==3.0.1
django-taggit==0.22.0
django-treebeard==4.1.0
djangorestframework==3.5.3
html5lib==0.9999999
lxml==3.7.3
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
pytz==2016.10
requests==2.13.0
six==1.10.0
Unidecode==0.4.20
wagtail==1.9
Willow==0.4
